Added the parts for the rocket
I added all the preliminary parts to the slider-item in order to get something in but now I have to adjust the width and size into the img-sliders in order for it to work.
also added more slider for the seocnd carousel and tried changing the wieghts of the rocket pieces to the accurate wieghts.